Fixed peak finder, working on fixing trim algorithm. Error in peak vector in powerWave.cpp, the peak values aren't appearing inside trim(), may be a pass by refernce error?

+++ v4-prototype/src/powerWave.h
@@ -28,7 +28,7 @@ class powerWave
     float getHarmonic(int i);
     float getHarmonicAngle(int i);
     size_t getNumHarmoics();
-    void trimData(std::vector<float>& data, std::vector<unsigned long>& peaks);
+    void trimData(std::vector<float>& data, std::vector<int>& peaks);
 
     size_t _errVal;
   protected:
